% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_aggregation.R
\name{gtf_predict}
\alias{gtf_predict}
\title{Compute gene-level p-values from Pelican site-wise p-values.}
\usage{
gtf_predict(
  dataset,
  ali_col = alignment,
  pval_col = aagtr_pval,
  naa_col = naa,
  k = 5,
  n = 10000
)
}
\arguments{
\item{dataset}{a tibble or dataframe that contains site p-values for a collection of genes}

\item{ali_col}{column of gene identifier}

\item{pval_col}{column of site p-values}

\item{naa_col}{column that contains the number of distinct amino acid at each site}

\item{k}{the number of best ranking p-values to consider at each gene}

\item{n}{the sample size to use in the empirical null distribution}
}
\description{
Computation of gene p-values is performed using the so-called
"Genewise Truncated Fisher" method.
It consists in compute a Fisher score on the best ranking p-values within each gene,
and use an empirical distribution of this score to calculate a p-value
for the corresponding gene.
}
\examples{
site_pvals = read_tsv("/path/to/pelican/output/all_sites.tsv")
# no progress bar
gene_predictions = site_pvals \%>\% gtf_predict()

# with progress bar
handlers("progress") # pick your favorite one, see `progressr` vignette
with_progress({
  gene_predictions = site_pvals \%>\% gtf_predict()
})

}
