% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfisher-package.R
\docType{package}
\name{gtfisher-package}
\alias{gtfisher}
\alias{gtfisher-package}
\title{gtfisher: Groupwise Truncated Fisher Score}
\description{
Implements the Groupwise Truncated Fisher (GTF) score to compute group-level \ p-values from independant p-values. \ It is designed as an alternative to Fisher's method, that does not suffer as much \ from the non uniformity of p-values under the null hypothesis. \ The API is designed to integrate with the package `dplyr`, \ and optionally with `progressr`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lsdch/gtfisher}
  \item Report bugs at \url{https://github.com/lsdch/gtfisher/issues}
}

}
\author{
\strong{Maintainer}: Louis Duchemin \email{ls.duchemin@gmail.com} (\href{https://orcid.org/0000-0002-7984-9554}{ORCID})

Authors:
\itemize{
  \item Philippe Veber
}

}
\keyword{internal}
