% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtf_aggregation.R
\name{gtf_predict}
\alias{gtf_predict}
\title{Compute aggregated p-value per group of p-values resulting from independent tests.}
\usage{
gtf_predict(dataset, group_col, pval_col, ..., k = 5, n = 10000)
}
\arguments{
\item{dataset}{a tibble or dataframe}

\item{group_col}{column of the grouping variable}

\item{pval_col}{column of p-values}

\item{...}{filtering condition as in dplyr::filter}

\item{k}{the number of best ranking p-values to consider in each group}

\item{n}{the sample size to use in the empirical null distribution}
}
\description{
Computation of aggregated p-values is performed using the so-called
"Groupwise Truncated Fisher" method.
It consists in compute a Fisher score on the best ranking p-values within each group,
then use an empirical distribution of this score to calculate a p-value
for the corresponding group.
}
\examples{
# no progress bar
group_pvals = gtf_predict(pvals, group, p)

# with progress bar
handlers("progress") # pick your favorite one, see `progressr` vignette
with_progress({
  group_pvals = gtf_predict(pvals, group, p)
})

}
